セッション切れ対応を追加

+++ share/src/employee/editFree.php
@@ -18,11 +18,6 @@ session_start();
 
 include('../../conf/config.php');
 $weekJa = array("日", "月", "火", "水", "木", "金", "土");
-for ($i = 0; $i < count($_GET['editFree']); $i++) {
-    list($empid[$i], $time[$i], $date[$i], $weekNum[$i]) = explode(":", $_GET['editFree'][$i]);
-    // timeフォーマット
-    $time[$i] = substr_replace($time[$i], ':', 2, 0) . ":00";
-}
 
 //データベース接続
 try {
@@ -31,33 +26,40 @@ try {
     $msg = $e->getMessage();
 }
 
-
-$sql = "SELECT id FROM rsvdb WHERE ";
-
-$arySql1 = array();
-for ($i = 0; $i < count($_GET['editFree']); $i++) {
-    $arySql1[$i] = '(empid = :empid' . $i . ' AND rsvdate = :rsvdate' . $i . ' AND rsvtime = :rsvtime' . $i . ')';
-}
-$sql .= implode(' OR ', $arySql1);
-
-//bind処理
-$stmt = $dbh->prepare($sql);
-for ($i = 0; $i < count($_GET['editFree']); $i++) {
-    $stmt->bindValue(':empid' . $i, $empid[$i]);
-    $stmt->bindValue(':rsvdate' . $i, date('Y') . '-' . str_replace('/', '-', $date[$i]));
-    $stmt->bindValue(':rsvtime' . $i, $time[$i]);
-}
-
-$stmt->execute();
-$deldata = $stmt->fetchAll(PDO::FETCH_ASSOC);
-
-if (!empty($deldata)) {
-    for ($i = 0; $i < count($deldata); $i++) {
-        $delid[$i] = $deldata[$i]['id'];
+if (!empty($_GET['editFree'])) {
+    for ($i = 0; $i < count($_GET['editFree']); $i++) {
+        list($empid[$i], $time[$i], $date[$i], $weekNum[$i]) = explode(":", $_GET['editFree'][$i]);
+        // timeフォーマット
+        $time[$i] = substr_replace($time[$i], ':', 2, 0) . ":00";
     }
-    $sql = "DELETE FROM rsvdb WHERE id IN (" . implode(',', $delid) . ")";
+
+    $sql = "SELECT id FROM rsvdb WHERE ";
+
+    $arySql1 = array();
+    for ($i = 0; $i < count($_GET['editFree']); $i++) {
+        $arySql1[$i] = '(empid = :empid' . $i . ' AND rsvdate = :rsvdate' . $i . ' AND rsvtime = :rsvtime' . $i . ')';
+    }
+    $sql .= implode(' OR ', $arySql1);
+
+    //bind処理
     $stmt = $dbh->prepare($sql);
+    for ($i = 0; $i < count($_GET['editFree']); $i++) {
+        $stmt->bindValue(':empid' . $i, $empid[$i]);
+        $stmt->bindValue(':rsvdate' . $i, date('Y') . '-' . str_replace('/', '-', $date[$i]));
+        $stmt->bindValue(':rsvtime' . $i, $time[$i]);
+    }
+
     $stmt->execute();
+    $deldata = $stmt->fetchAll(PDO::FETCH_ASSOC);
+
+    if (!empty($deldata)) {
+        for ($i = 0; $i < count($deldata); $i++) {
+            $delid[$i] = $deldata[$i]['id'];
+        }
+        $sql = "DELETE FROM rsvdb WHERE id IN (" . implode(',', $delid) . ")";
+        $stmt = $dbh->prepare($sql);
+        $stmt->execute();
+    }
 }
 ?>
 
@@ -70,20 +72,26 @@ if (!empty($deldata)) {
     </header>
 
     <main>
-        <div class="container">
-            <h1>COMPLETE</h1>
-            <div class="btn">
-                <form action="./editFree_form.php" method="get">
-                    <input type="hidden" name="empid" value="<?php echo $empid[0]; ?>">
-                    <input type="hidden" name="week" value="0">
-                    <button type="submit" id="register">追加で削除する</button>
-                </form>
-                <form action="./empmypage.php" method="get">
-                    <input type="hidden" name="week" value="0">
-                    <input type="hidden" name="empid" value="<?php echo $empid[0]; ?>">
-                    <button id="backHome" onclick="location.href='./empmypage.php'">ホームへ戻る</button>
-                </form>
+        <?php if (!empty($_GET['editFree']) && !empty($_SESSION['eid'])) { ?>
+            <div class="container">
+                <h1>COMPLETE</h1>
+                <div class="btn">
+                    <form action="./editFree_form.php" method="get">
+                        <input type="hidden" name="empid" value="<?php echo $empid[0]; ?>">
+                        <input type="hidden" name="week" value="0">
+                        <button type="submit" id="register">追加で削除する</button>
+                    </form>
+                    <form action="./empmypage.php" method="get">
+                        <input type="hidden" name="week" value="0">
+                        <input type="hidden" name="empid" value="<?php echo $empid[0]; ?>">
+                        <button id="backHome" onclick="location.href='./empmypage.php'">ホームへ戻る</button>
+                    </form>
+                </div>
             </div>
-        </div>
+        <?php } else { ?>
+            <h1>セッションが切れました</h1>
+            <h2>再ログインしてください</h2>
+            <a href="./emplogin_form.php">ログイン</a>
+        <?php } ?>
     </main>
 </body>
